pyproject.toml
[project]
name = "llm-ielts-poc"
version = "0.1.0"
description = "PoC: IELTS Task1/Task2 scorer with evidence-first JSON outputs"
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
  "fastapi==0.112.2",
  "uvicorn[standard]==0.30.6",
  "pydantic==2.8.2",
  "pydantic-settings==2.4.0",
  "httpx==0.27.0",
  "jsonschema==4.23.0",
  "python-dotenv==1.0.1",
  "ulid-py==1.1.0",
  "typing-extensions>=4.12.2",
  "openai==1.54.3",
  "pandas==2.2.2",
  "numpy==1.26.4",
  "scikit-learn==1.5.1",
  "scipy==1.13.1",
  "datasets==2.21.0",
]

[project.optional-dependencies]
dev = [
  "ruff==0.6.8",
  "pytest==8.3.2",
  "pytest-cov==5.0.0",
  "mypy==1.11.2",
]

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I", "B", "UP"]
ignore = []

[tool.ruff.isort]
known-first-party = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
pythonpath = ["src"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_unused_ignores = true
pretty = true
